Computing with Cells and Atoms (Calude & Paun 2000)
Title: "Computing with Cells and Atoms: An Introduction to Quantum, DNA and Membrane Computing"
via: https://www.cs.auckland.ac.nz/~cristian/cca.pdf
pages:105..108
Year:2000
1.3 A General Computing Framework
The aim of this section is to “define”, in a general, somewhat formal, but
not in a very precise way, the notions of computation and computing device/model.
In fact, we are not looking for a comprehensive definition, but
for a comprehensive framework for an activity which can be called a computation.
We do not address here semiotic or philosophical questions (can
a computation be a natural process, or is it a specific human artifact? is a
process a computation per se, or only for some observer? and so on), but
we step into an operational direction: we want to systematize the notions
we can encounter when devising a computing machinery, in a comprehensive
manner which is also general enough. Otherwise stated, we do not adopt a
minimali